Consolidated the process test fixtures; moved the share modal test to the modeler test file; moved the mockClipboard setup to the base page setup
